cloud-functions/requirements.txt
# Cloud Functions Dependencies
functions-framework==3.4.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.13.0
google-cloud-speech==2.21.0
google-generativeai==0.3.2
firebase-admin==6.2.0
